+++ backend/model/heater.go
@@ -3,7 +3,7 @@ package model
 import "github.com/trichner/berryhunter/backend/phy"
 
 type HeatRadiator struct {
-	HeatPerTick float32
+	HeatPerTick uint32
 	Radius      float32
 	Body        phy.DynamicCollider
 }
